Move ResourceHints component so it is no longer nested under Document
