+++ src/app/components/ResourceHints/index.jsx
